refactor(package): extract verification code
